Mobile-friendly view of filters, components separated so that they can be reused.
